Refactor modal into export component
